Move VmSlice to transaction context
