+++ program-runtime/src/memory.rs
@@ -2,6 +2,7 @@
 
 use {
     solana_sbpf::memory_region::{AccessType, MemoryMapping},
+    solana_transaction_context::vm_slice::VmSlice,
     std::{mem::align_of, slice::from_raw_parts_mut},
 };
 
@@ -126,4 +127,25 @@ pub fn translate_slice_mut_for_cpi<'a, T>(
         T,
         check_aligned,
     )
+}
+
+pub trait TranslateSlice {
+    type Data;
+    /// Returns a slice using a mapped physical address
+    fn translate<'a>(
+        &self,
+        memory_mapping: &'a MemoryMapping,
+        check_aligned: bool,
+    ) -> Result<&'a [Self::Data], Box<dyn std::error::Error>>;
+}
+
+impl<T> TranslateSlice for VmSlice<T> {
+    type Data = T;
+    fn translate<'a>(
+        &self,
+        memory_mapping: &'a MemoryMapping,
+        check_aligned: bool,
+    ) -> Result<&'a [T], Box<dyn std::error::Error>> {
+        translate_slice::<T>(memory_mapping, self.ptr(), self.len(), check_aligned)
+    }
 }
